+++ include/minishell.h
@@ -6,7 +6,7 @@
 /*   By: fschuber <fschuber@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/30 11:30:11 by fschuber          #+#    #+#             */
-/*   Updated: 2024/04/26 17:46:25 by fschuber         ###   ########.fr       */
+/*   Updated: 2024/04/26 19:01:59 by fschuber         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -42,8 +42,8 @@
 
 // will output detailed logging if set to 1, and normal logging if 0
 #define VERBOSE 0
-// will make everything work with tester provided executable is called minishell
-#define DEBUG 1
+// will not print logo if set to 1
+#define DEBUG 0
 
 // ----- TOKEN TYPES
 
@@ -97,7 +97,6 @@ typedef struct s_data
 	int				exit_status;
 	char			**envcp;
 	t_list			*gc;
-	t_pid_list		*pid_list;
 }	t_data;
 
 typedef struct s_cmd_path
@@ -164,7 +163,6 @@ int					check_brace_errors(t_list *tok, t_tok *token, \
 int					check_first_token(t_list *tok);
 int					check_last_token(t_tok *token);
 int					check_braces(int brace_opened);
-int					process_tok_list(t_list **tok, int *brace_opened);
 void				throw_syntax_error(char *error);
 
 // --- 2-expander
@@ -179,9 +177,7 @@ void				handle_dollar_question_expansion(t_exp *exp, t_data *data);
 void				handle_tilde_expansion(t_exp *exp, t_data *data);
 // wildcard
 void				handle_wildcard_expansion(t_exp *exp, t_data *data);
-char				*list_matching_files(char *pattern);
 char				*get_pattern(const char *str, int index, t_data *sh);
-char				*get_rid_of_quotes_wildcard(char *str);
 // expand util
 int					find_closing_quote(const char *str, int *i);
 char				*get_envcp(const char *var_name, t_data *sh);
@@ -210,8 +206,8 @@ int					execute_node(t_node *node, t_data *data, \
 										t_pid_list **pid_list);
 void				child_process_exit(t_data *data, int exitcode);
 // operators
-int					logical_and(t_node *node, t_data *sh);
-int					logical_or(t_node *node, t_data *sh);
+int					logical_and(t_node *node, t_data *sh, t_pid_list **pidlist);
+int					logical_or(t_node *node, t_data *sh, t_pid_list **pidlist);
 int					redirect(t_node *node, t_data *data);
 int					heredoc(t_node *node, t_data	*sh);
 void				setup_pipe(t_node *node);